requirements.txt
apiclient==1.0.4
attrs==23.1.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
exceptiongroup==1.1.3
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
iniconfig==2.0.0
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
pluggy==1.3.0
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
PySocks==1.7.1
pytest==7.4.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.23.1
trio-websocket==0.11.1
undetected-chromedriver==3.5.3
uritemplate==4.1.1
urllib3==2.0.7
websockets==12.0
wsproto==1.2.0
